#37 Typo in one of the return types for toFloat

+++ jvm/src/main/scala-2.10/scalajson.ast/unsafe/JValue.scala
@@ -85,7 +85,7 @@ final case class JNumber(value: String) extends JValue {
 
   def toDouble: Double = scalajson.ast.toDouble(value)
 
-  def toFloat: Double = scalajson.ast.toFloat(value)
+  def toFloat: Float = scalajson.ast.toFloat(value)
 
   def toBigDecimal: Option[BigDecimal] = scalajson.ast.toBigDecimal(value)
 }
